now! side chain and main chain are simultaneously running ...
found why it is happening!!:nerd_face:
for future reference:
Onet have a function called: Dispatch() that what it does is basically listening for messages and it is a medium between protocols and network layer.
see protocol.go in Onet, protocolinstance interface: “Dispatch is called at the beginning by Onet for listening on the channels”
So, in order to have two protocols running at the same time (and listening to the message exchanged in both of them) I changed this dispatch function a little bit - because Dispatch was called initially creating blscosi protocol, so I couldn’t listen for message from this protocol if it wasn’t yet created.
I added another Dispatch frunction named: DispatchProtocol() and added it to BaseDfs protocol, listening for both types of message (from both protocols). and then called it in runsimul.go after both protocols have been created.
:bulb:Turns out that this Dispatch function had not only been called in protocol creation time (to listen for messages) , but it is called while handling messages in overlay.go - overlay is a layer between network and protocol - in function: TransmitMsg()
so when the messages were sent, the empty dispatch function were called to handle the message from the receiver side (I don’t want to go into why it is working like this).
:high_brightness: problem solved: now there is two dispatch functions in basedfs.go, one for baseDFS protocol and one for interacting with sideChain BlsCosi protocol.