working on the committee
side chain and main chain are working together, and the committee is dynamically changing.
the problem is probably in when we have repetitive nodes in the committee that cause blscosi protocol to hang over (no err msg and no progress)
and also after changing the committee in different point somewhere its length gets one less than required.