What happens (98% sure!) in distributed simulation is the description::
1- simul.start()
2- (simul.go): start()
  3- (build.go): startBuild()
  4- platform.ReadRunFile()
  5- deploy.configure()
  6- RunTests()
    7- deploy.build()
       8- platform.build()
  9- runTest()
    10- deployP.deploy // get list of distributed server addresses and copy over the generated exe and config files
    11- deployP.start()
        12- ssh ./users
in users.go the function. main is called which in turn will get to ssh ./simul and the big discovery!!!!! is that this time the "simul" flag will be not empty which will mean that when Start() is called {it is not called now, it is located in simul.go in main package in main function, it must be called!} this time it will be initialized with non empty "simul" flag and this means it gets to call "platform.Simulate" which starts the protocol. :|